refactor: decompose scannerService into focused modules
Split the 397-line scannerService.ts into 4 modules:
- filesystemCollector.ts (145 lines): file traversal, root sorting, state collection
- changeDetector.ts (46 lines): pure change classification logic
- trackBuilder.ts (120 lines): probing, metadata override, merging
- scannerService.ts (85 lines): orchestration only

No behavior changes. All existing tests pass unchanged.
